Add Markdown parser interface
