+++ src/main/java/com/elovirta/dita/markdown/Schema.java
@@ -1,11 +1,10 @@
 package com.elovirta.dita.markdown;
 
-import com.vladsch.flexmark.util.data.MutableDataSet;
-
 import java.net.URI;
 import java.util.Set;
 
 public interface Schema {
-    Set<URI> getUri();
-    MutableDataSet getOptions();
+    Set<URI> getScheme();
+
+    MarkdownParser createMarkdownParser();
 }
